+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #93 LockHandlers to be contributed via extenders to a DI Map
 * #94 Bump spring-core from 5.3.9 to 5.3.14
 * #95 Failure in JobListener must fail the job
+* #96 Enhanced API for job parameters
 
 ## 2.0.M1